Use the capabilities getmap / gettile url as the "base" for templates, fall back to using the URL of the capabilities document itself.
Update MapMLBaseProxyTest.java to check the cascaded request against the
request URL from the capabilities document OR if that is null, use the
URL OF the capabilities document as the base.

Update assertCascading in MapMLWMTSProxyTest.java to compare against
resource request URL from capabilities document OR if that is null,
use the URL OF the capabilities document as the base.